backend/requirements.txt
Django==3.2.16
djangorestframework==3.13.1
django-cors-headers==3.8.0
uWSGI==2.0.20
django-webpack-loader==1.7.0
requests
debugpy